+++ lib/src/api/entities/session/partial_session.dart
@@ -1,0 +1,10 @@
+import '../../../../restrr.dart';
+
+abstract class PartialSession extends RestrrEntity {
+  String? get name;
+  DateTime get createdAt;
+  DateTime get expiresAt;
+  User get user;
+
+  Future<bool> delete();
+}