+++ lib/src/api/exceptions/restrr_exception.dart
@@ -1,0 +1,8 @@
+abstract class RestrrException implements Exception {
+  final String? message;
+
+  RestrrException(this.message);
+
+  @override
+  String toString() => '${runtimeType.toString()}: $message';
+}