+++ lib/src/internal/entities/session/session_impl.dart
@@ -1,0 +1,18 @@
+import 'package:restrr/src/internal/entities/session/partial_session_impl.dart';
+
+import '../../../../restrr.dart';
+
+class SessionImpl extends PartialSessionImpl implements Session {
+  @override
+  final String token;
+
+  const SessionImpl({
+    required super.api,
+    required super.id,
+    required super.name,
+    required super.createdAt,
+    required super.expiresAt,
+    required super.user,
+    required this.token
+  });
+}